requirements.txt
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.7.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.8.2
build==1.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
docstring-parser==0.15
docx==0.2.4
ebcdic==1.1.1
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0
filelock==3.13.4
firecrawl-py==0.0.8
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
groq==0.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.1
idna==3.10
iniconfig==2.1.0
instructor==1.0.2
ipykernel==6.29.4
ipython==8.23.0
ISO8583==1.4
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
langchain==0.1.16
langchain-anthropic==0.3.7
langchain-community==0.0.33
langchain-core==0.3.35
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langgraph==0.2.71
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy==1.8.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.2
orjson==3.10.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
pillow_heif==0.21.0
pinecone-client==3.2.2
pip-tools==7.4.1
pipenv==2023.12.1
platformdirs==4.2.0
playwright==1.42.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyee==11.0.1
Pygments==2.17.2
PyMuPDF==1.24.14
pyparsing==3.2.0
pypdfium2==4.30.1
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==8.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.0
rapidfuzz==3.3.1
reducto==1.0.3
reductoai==0.1.0
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
retell-sdk==4.0.0
rich==13.9.4
safetensors==0.4.5
scipy==1.14.1
selenium==4.19.0
sendgrid==6.11.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
spandrel==0.4.0
SQLAlchemy==2.0.29
stack-data==0.6.3
starkbank-ecdsa==2.2.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.6.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchsde==0.2.6
torchvision==0.19.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
trampoline==0.1.2
transformers==4.45.1
trieve_client_py @ git+https://github.com/devflowinc/trieve_client_py@7fde5e8d745c989555aa25544d2a30690f34704b
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zstandard==0.23.0
